Refactor JSON schema handling and serialization in ReflectorNet
- Introduced a new JsonSchema class to manage JSON schema generation and processing.
- Added methods to Reflector for obtaining schemas for types and method arguments.
- Enhanced error handling in schema generation to provide clearer feedback.
- Consolidated JSON serialization logic into a dedicated JsonSerializer class.
- Removed deprecated JsonUtils.Schema and JsonUtils classes, streamlining the codebase.
- Added constants for JSON schema properties to improve maintainability.
- Updated Reflector.Error to utilize new schema generation methods.
- Improved handling of nullable types in schema generation.
- Added utility methods for unstringifying JSON elements.
- Enhanced StringUtils with a method to check for null or whitespace values.
