+++ ReflectorNet/src/Utils/Json/JsonSerializer.Consts.cs
@@ -1,0 +1,9 @@
+namespace com.IvanMurzak.ReflectorNet.Utils
+{
+    public partial class JsonSerializer
+    {
+        public const string Null = "null";
+        public const string EmptyJsonObject = "{}";
+        public const string EmptyJsonArray = "[]";
+    }
+}
